SLING-10135 cleanup and code coverage
Update to parent 40 to improve jacoco config
Enable jacoco.command for the integration tests
convert tabs to spaces and other whitespace cleanup
